moved jetty rest tests to appropriate package
