+++ commons-jetty/src/test/scala/com/avsystem/commons/jetty/rest/HttpRestCallTest.scala
@@ -1,7 +1,6 @@
 package com.avsystem.commons
-package jetty.rpc
+package jetty.rest
 
-import com.avsystem.commons.jetty.rest.{RestClient, RestServlet}
 import com.avsystem.commons.rest.AbstractRestCallTest
 import com.avsystem.commons.rest.RawRest.HandleRequest
 import org.eclipse.jetty.server.Server
