+++ commons-jetty/src/test/scala/com/avsystem/commons/jetty/rest/UsesHttpServer.scala
@@ -1,5 +1,5 @@
 package com.avsystem.commons
-package jetty.rpc
+package jetty.rest
 
 import org.eclipse.jetty.server.Server
 import org.scalatest.{BeforeAndAfterAll, Suite}
